// src/app/models/environmental-ethical.model.ts
export interface EnvironmentalEthicalData {
    environmentalPolicy: boolean;
    environmentalPolicyPublic: boolean;
    climateChangeStrategy: boolean;
    carbonNeutralByYear: number;
    emissionsReductionTarget: number;
    carbonOffsetProgram: boolean;
    renewableEnergyPercentage: number;
    wasteReductionTarget: number;
    circularEconomyInitiatives: boolean;
    waterConservationProgram: boolean;
    environmentalImpactDescription: string;
  }